bot/requirements.txt
psycopg2-binary==2.9.6
aiogram==2.25.1
python-dotenv==1.0.0
Pillow>=10